requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0.4
django-allauth==0.62.1
django-cache-url==3.4.5
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-jalali-date==1.1.0
django-js-asset==2.2.0
django-rosetta==0.10.0
environs==11.0.0
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
marshmallow==3.21.2
packaging==24.0
pillow==10.3.0
polib==1.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
